Revert "add gst-plugins-ugly"
This reverts commit 833554a9df8d5db8908f33c9483eaab261979b32.

+++ gst-plugins-ugly/ext/sidplay/gstsiddec.h
@@ -58,7 +58,7 @@ struct _GstSidDec {
   sidTune       *tune;
   emuConfig     *config;
 
-  guint         blocksize;
+  gulong         blocksize;
 };
 
 struct _GstSidDecClass {
